Broke table into multiple components
