+++ client/src/components/toolbox/inventory/products/create/CreateProducts.js
@@ -123,6 +123,7 @@ export default class CreateProducts extends Component {
 
     render() {
         const { table } = this.state
+        // console.log('createProducts rendered')
 
         return (
             <Segment inverted style={{ background: '#252525', minHeight: '100vh' }}>
